requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
requests>=2.31.0
google-generativeai>=0.3.2
aiohttp>=3.8.6
googletrans==4.0.0-rc1
openpyxl>=3.1.2
python-docx>=1.1.0
python-pptx>=0.6.21
PyMuPDF>=1.23.9
pytesseract>=0.3.10
Pillow>=10.2.0
aiofiles>=23.2.1

